+++ R/tree_information.R
@@ -8,6 +8,11 @@
 #' @param x A tree of class `phylo`, a list of trees, or a `multiPhylo` object.
 #' @param p A vector of probabilities corresponding t oeach split in `x`. 
 #' Specify `TRUE` to calculate this vector using the node labels of each tree.
+#' @param sum Logical: if `TRUE`, sum the information content of each split to
+#' provide the total splitwise information content of the tree.
+#' 
+#' @return `SplitwiseInfo()` returns the splitwise information content of the
+#' tree (or of each split in turn, if `sum = FALSE`), in bits.
 #' 
 #' @family information functions
 #' 
@@ -24,8 +29,9 @@
 #' SplitwiseInfo(tree, TRUE)
 #' @template MRS
 #' @export
-SplitwiseInfo <- function (x, p = NULL) UseMethod('SplitwiseInfo')
-
+SplitwiseInfo <- function (x, p = NULL, sum = TRUE) {
+  UseMethod('SplitwiseInfo')
+}
 
 .GetPFromLabels <- function (tree, p, splits = as.Splits(tree)) {
   if (length(p) == 1L) { # length(NULL) == 0
@@ -54,8 +60,8 @@ SplitwiseInfo.phylo <- function (x, p = NULL) {
 }
   
 #' @export
-SplitwiseInfo.multiPhylo <- function (x, p = NULL) {
-  vapply(as.Splits(x), SplitwiseInfo, double(1L), p)
+SplitwiseInfo.multiPhylo <- function (x, p = NULL, sum = TRUE) {
+  vapply(as.Splits(x), SplitwiseInfo, double(1L), p, sum)
 }
 
 #' @export
@@ -63,7 +69,7 @@ SplitwiseInfo.list <- SplitwiseInfo.multiPhylo
 
 #' @importFrom TreeTools Log2Rooted.int Log2Unrooted.int TipsInSplits
 #' @export
-SplitwiseInfo.Splits <- function(x, p = NULL) {
+SplitwiseInfo.Splits <- function(x, p = NULL, sum = TRUE) {
   nTip <- attr(x, 'nTip')
   inSplit <- TipsInSplits(x)
   
@@ -88,14 +94,21 @@ SplitwiseInfo.Splits <- function(x, p = NULL) {
     l2nInconsistent <- l2pInconsistent + l2n
     
     # Return:
-    sum(l2n * length(inSplit),
-        p * (log2(p) - l2nConsistent),
-        q * (log2(q) - l2nInconsistent))
+    if (sum) {
+      sum(l2n * length(inSplit),
+          p * (log2(p) - l2nConsistent),
+          q * (log2(q) - l2nInconsistent))
+    } else {
+      ret <- l2n + (p * (log2(p) - l2nConsistent)) 
+      ret[names(l2nInconsistent)] <- ret[names(l2nInconsistent)] +
+        (q * (log2(q) - l2nInconsistent))
+      ret
+    }
   }
 }
 
 #' @export
-SplitwiseInfo.NULL <- function (x, p = NULL) 0
+SplitwiseInfo.NULL <- function (x, p = NULL, sum = TRUE) 0
 
 #' @rdname SplitwiseInfo
 #' @param check.tips Logical specifying whether to renumber leaves such that
@@ -222,23 +235,27 @@ ConsensusInfo <- function (trees, info = 'phylogenetic', check.tips = TRUE) {
 #' @family information functions
 #' @importFrom TreeTools as.Splits
 #' @export
-ClusteringEntropy <- function (x, p = NULL) UseMethod("ClusteringEntropy")
-
-#' @rdname ClusteringEntropy
-#' @export
-ClusteringInfo <- function (x, p = NULL) UseMethod("ClusteringInfo")
-
-#' @rdname ClusteringEntropy
-#' @export
-ClusteringEntropy.phylo <- function (x, p = NULL) {
-  splits <- as.Splits(x)
-  ClusteringEntropy.Splits(splits, p = .GetPFromLabels(x, p, splits))
+ClusteringEntropy <- function (x, p = NULL, sum = TRUE) {
+  UseMethod("ClusteringEntropy")
 }
 
 #' @rdname ClusteringEntropy
 #' @export
-ClusteringEntropy.list <- function (x, p = NULL)
-    vapply(as.Splits(x), ClusteringEntropy.Splits, double(1L), p = p)
+ClusteringInfo <- function (x, p = NULL, sum = TRUE) {
+  UseMethod("ClusteringInfo")
+}
+
+#' @rdname ClusteringEntropy
+#' @export
+ClusteringEntropy.phylo <- function (x, p = NULL, sum = TRUE) {
+  splits <- as.Splits(x)
+  ClusteringEntropy.Splits(splits, p = .GetPFromLabels(x, p, splits), sum)
+}
+
+#' @rdname ClusteringEntropy
+#' @export
+ClusteringEntropy.list <- function (x, p = NULL, sum = TRUE)
+    vapply(as.Splits(x), ClusteringEntropy.Splits, double(1L), p = p, sum)
 
 #' @rdname ClusteringEntropy
 #' @export
@@ -246,7 +263,7 @@ ClusteringEntropy.multiPhylo <- ClusteringEntropy.list
 
 #' @rdname ClusteringEntropy
 #' @export
-ClusteringEntropy.Splits <- function (x, p = NULL) {
+ClusteringEntropy.Splits <- function (x, p = NULL, sum = TRUE) {
   nLeaves <- attr(x, 'nTip')
   inSplit <- TipsInSplits(x)
   splitP <- rbind(inSplit, nLeaves - inSplit, deparse.level = 0L) / nLeaves
@@ -254,24 +271,25 @@ ClusteringEntropy.Splits <- function (x, p = NULL) {
     p <- 1L
   }
   
+  ret <- p * apply(splitP, 2, Entropy)
   # Return:
-  sum(p * apply(splitP, 2, Entropy))
+  if (sum) sum(ret) else ret
 }
 
 #' @export
-ClusteringEntropy.NULL <- function (x, p = NULL) NULL
+ClusteringEntropy.NULL <- function (x, p = NULL, sum = TRUE) NULL
 
 #' @rdname ClusteringEntropy
 #' @export
-ClusteringInfo.phylo <- function (x, p = NULL) {
+ClusteringInfo.phylo <- function (x, p = NULL, sum = TRUE) {
   splits <- as.Splits(x)
-  ClusteringInfo.Splits(splits, p = .GetPFromLabels(x, p, splits))
+  ClusteringInfo.Splits(splits, p = .GetPFromLabels(x, p, splits), sum)
 }
 
 #' @rdname ClusteringEntropy
 #' @export
-ClusteringInfo.list <- function (x, p = NULL)
-    vapply(as.Splits(x), ClusteringInfo.Splits, double(1L), p = p)
+ClusteringInfo.list <- function (x, p = NULL, sum = TRUE)
+    vapply(as.Splits(x), ClusteringInfo.Splits, double(1L), p = p, sum)
 
 #' @rdname ClusteringEntropy
 #' @export
@@ -279,17 +297,10 @@ ClusteringInfo.multiPhylo <- ClusteringInfo.list
 
 #' @rdname ClusteringEntropy
 #' @export
-ClusteringInfo.Splits <- function (x, p = NULL) {
+ClusteringInfo.Splits <- function (x, p = NULL, sum = TRUE) {
   nLeaves <- attr(x, 'nTip')
-  inSplit <- TipsInSplits(x)
-  splitP <- rbind(inSplit, nLeaves - inSplit, deparse.level = 0L) / nLeaves
-  if (is.null(p)) {
-    p <- 1L
-  }
-  
-  # Return:
-  sum(p * apply(splitP, 2, Entropy)) * nLeaves
+  ClusteringEntropy.Splits(x, p, sum) * nLeaves
 }
 
 #' @export
-ClusteringInfo.NULL <- function (x) NULL
+ClusteringInfo.NULL <- function (x, p = NULL, sum = TRUE) NULL