+++ NEWS.md
@@ -1,9 +1,10 @@
 # TreeDist 2.0.3.9004 (development)
 
-- `SplitwiseInfo()` and `ClusteringInfo()` gain a `p` parameter to reflect the
-  reduced information content of splits with lower support values.
 - `cons_phylo_info()` quickly calculates the information content of the
   consensus of a set of trees.
+- `SplitwiseInfo()` and `ClusteringInfo()` gain a `p` parameter to reflect the
+  reduced information content of splits with lower support values, and a `sum`
+  parameter to allow return of individual split information content.
 
 # TreeDist 2.0.3.9002 (development)